pyproject.toml
[project]
name = "shmup"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    'pygame-ce',
]
